+++ Source/WebCore/editing/cocoa/EditorCocoa.mm
@@ -52,8 +52,8 @@
 #import "HTMLPictureElement.h"
 #import "HTMLSourceElement.h"
 #import "HTMLSpanElement.h"
-#import "ImageBufferUtilitiesCG.h"
 #import "ImageOverlay.h"
+#import "ImageUtilities.h"
 #import "LegacyNSPasteboardTypes.h"
 #import "LegacyWebArchive.h"
 #import "LocalFrame.h"
