+++ Source/WebCore/platform/graphics/ImageBuffer.cpp
@@ -36,6 +36,7 @@
 #include "HostWindow.h"
 #include "ImageBufferDisplayListBackend.h"
 #include "ImageBufferPlatformBackend.h"
+#include "ImageUtilities.h"
 #include "MIMETypeRegistry.h"
 #include "ProcessCapabilities.h"
 #include "TransparencyLayerContextSwitcher.h"
@@ -45,17 +46,11 @@
 
 #if USE(CG)
 #include "ImageBufferCGPDFDocumentBackend.h"
-#include "ImageBufferUtilitiesCG.h"
-#endif
-
-#if USE(CAIRO)
-#include "ImageBufferUtilitiesCairo.h"
 #endif
 
 #if USE(SKIA)
 #include "GLContext.h"
 #include "ImageBufferSkiaAcceleratedBackend.h"
-#include "ImageBufferUtilitiesSkia.h"
 #include "PlatformDisplay.h"
 
 WTF_IGNORE_WARNINGS_IN_THIRD_PARTY_CODE_BEGIN
@@ -276,16 +271,6 @@ static RefPtr<NativeImage> copyImageBufferToNativeImage(Ref<ImageBuffer> source,
     return ImageBuffer::sinkIntoNativeImage(WTF::move(copyBuffer));
 }
 
-static RefPtr<NativeImage> copyImageBufferToOpaqueNativeImage(Ref<ImageBuffer> source, PreserveResolution preserveResolution)
-{
-    // Composite this ImageBuffer on top of opaque black, because JPEG does not have an alpha channel.
-    auto copyBuffer = copyImageBuffer(WTF::move(source), preserveResolution);
-    if (!copyBuffer)
-        return { };
-    // We composite the copy on top of black by drawing black under the copy.
-    copyBuffer->context().fillRect({ { }, copyBuffer->logicalSize() }, Color::black, CompositeOperator::DestinationOver);
-    return ImageBuffer::sinkIntoNativeImage(WTF::move(copyBuffer));
-}
 
 RefPtr<ImageBuffer> ImageBuffer::clone() const
 {
@@ -478,36 +463,6 @@ void ImageBuffer::transformToColorSpace(const DestinationColorSpace& newColorSpa
         backend->transformToColorSpace(newColorSpace);
         m_parameters.colorSpace = newColorSpace;
     }
-}
-
-String ImageBuffer::toDataURL(const String& mimeType, std::optional<double> quality, PreserveResolution preserveResolution) const
-{
-    return toDataURL(Ref { const_cast<ImageBuffer&>(*this) }, mimeType, quality, preserveResolution);
-}
-
-Vector<uint8_t> ImageBuffer::toData(const String& mimeType, std::optional<double> quality, PreserveResolution preserveResolution) const
-{
-    return toData(Ref { const_cast<ImageBuffer&>(*this) }, mimeType, quality, preserveResolution);
-}
-
-String ImageBuffer::toDataURL(Ref<ImageBuffer> source, const String& mimeType, std::optional<double> quality, PreserveResolution preserveResolution)
-{
-    auto encodedData = toData(WTF::move(source), mimeType, quality, preserveResolution);
-    if (encodedData.isEmpty())
-        return "data:,"_s;
-    return makeString("data:"_s, mimeType, ";base64,"_s, base64Encoded(encodedData));
-}
-
-Vector<uint8_t> ImageBuffer::toData(Ref<ImageBuffer> source, const String& mimeType, std::optional<double> quality, PreserveResolution preserveResolution)
-{
-    RefPtr<NativeImage> image = MIMETypeRegistry::isJPEGMIMEType(mimeType) ? copyImageBufferToOpaqueNativeImage(WTF::move(source), preserveResolution) : copyImageBufferToNativeImage(WTF::move(source), DontCopyBackingStore, preserveResolution);
-    if (!image)
-        return { };
-#if USE(SKIA)
-    return encodeData(*image, mimeType, quality);
-#elif USE(CG) || USE(CAIRO)
-    return encodeData(image->platformImage().get(), mimeType, quality);
-#endif
 }
 
 RefPtr<PixelBuffer> ImageBuffer::getPixelBuffer(const PixelBufferFormat& destinationFormat, const IntRect& sourceRect, const ImageBufferAllocator& allocator) const
